% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_fetch_manifest}
\alias{backend_fetch_manifest}
\title{Fetch and parse a YAML manifest from the backend}
\usage{
backend_fetch_manifest(backend, uri)
}
\arguments{
\item{backend}{A dsimaging_backend object.}

\item{uri}{Character; URI of the YAML file.}
}
\value{
Parsed YAML as a named list.
}
\description{
Fetch and parse a YAML manifest from the backend
}
