% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_list}
\alias{backend_list}
\title{List objects under a URI prefix}
\usage{
backend_list(backend, prefix)
}
\arguments{
\item{backend}{A dsimaging_backend object.}

\item{prefix}{Character; URI prefix.}
}
\value{
Character vector of full URIs.
}
\description{
List objects under a URI prefix
}
