% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_put_directory}
\alias{backend_put_directory}
\title{Upload a local directory to the backend (recursive)}
\usage{
backend_put_directory(backend, local_dir, uri_prefix)
}
\arguments{
\item{backend}{A dsimaging_backend object.}

\item{local_dir}{Character; local directory.}

\item{uri_prefix}{Character; destination URI prefix.}
}
\value{
Invisible uri_prefix.
}
\description{
Upload a local directory to the backend (recursive)
}
