% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{backend_put_file}
\alias{backend_put_file}
\title{Upload a local file to the backend}
\usage{
backend_put_file(backend, local_path, uri)
}
\arguments{
\item{backend}{A dsimaging_backend object.}

\item{local_path}{Character; local file to upload.}

\item{uri}{Character; destination URI.}
}
\value{
Invisible URI.
}
\description{
Upload a local file to the backend
}
