% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.assert_min_samples}
\alias{.assert_min_samples}
\title{Assert that a dataset has enough samples}
\usage{
.assert_min_samples(n_samples, context = "dataset")
}
\arguments{
\item{n_samples}{Integer; number of samples.}

\item{context}{Character; context for error message.}
}
\description{
Assert that a dataset has enough samples
}
\keyword{internal}
