% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{.asset_db_connect}
\alias{.asset_db_connect}
\title{Open the asset catalog database}
\usage{
.asset_db_connect()
}
\value{
A DBI connection.
}
\description{
Open the asset catalog database
}
\keyword{internal}
