% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{.asset_db_prepare_path}
\alias{.asset_db_prepare_path}
\title{Prepare asset DB directory and repair common permission drift.}
\usage{
.asset_db_prepare_path(db_path = .asset_db_path())
}
\description{
Prepare asset DB directory and repair common permission drift.
}
\keyword{internal}
