% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{.asset_db_repair_permissions}
\alias{.asset_db_repair_permissions}
\title{Repair permissions on the SQLite database sidecar files.}
\usage{
.asset_db_repair_permissions(db_path = .asset_db_path())
}
\description{
Repair permissions on the SQLite database sidecar files.
}
\keyword{internal}
