% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{.asset_find_by_hash}
\alias{.asset_find_by_hash}
\title{Find asset by derivation_hash (deduplication)}
\usage{
.asset_find_by_hash(
  db,
  dataset_id,
  derivation_hash,
  visibility = "global",
  collection_seal = NULL
)
}
\description{
Find asset by derivation_hash (deduplication)
}
\keyword{internal}
