% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{.asset_list_aliases}
\alias{.asset_list_aliases}
\title{List aliases for a dataset}
\usage{
.asset_list_aliases(db, dataset_id, collection_seal = NULL)
}
\description{
List aliases for a dataset
}
\keyword{internal}
