% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{.asset_resolve_alias}
\alias{.asset_resolve_alias}
\title{Resolve an alias to an asset_id}
\usage{
.asset_resolve_alias(db, dataset_id, alias, collection_seal = NULL)
}
\description{
Resolve an alias to an asset_id
}
\keyword{internal}
