% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentHashDS.R
\name{.content_hash_handle_minio_event}
\alias{.content_hash_handle_minio_event}
\title{Handle a MinIO/S3 bucket notification payload}
\usage{
.content_hash_handle_minio_event(payload)
}
\description{
Handle a MinIO/S3 bucket notification payload
}
\keyword{internal}
