% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_capabilities.R
\name{.dshpc_option_safe}
\alias{.dshpc_option_safe}
\title{Read dshpc option safely (dsImaging doesn't import dsHPC)}
\usage{
.dshpc_option_safe(name)
}
\description{
Read dshpc option safely (dsImaging doesn't import dsHPC)
}
\keyword{internal}
