% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contentHashDS.R
\name{.dsimagingstore_option}
\alias{.dsimagingstore_option}
\title{Resolve an option using the dsHPC-style option chain}
\usage{
.dsimagingstore_option(name, default = NULL)
}
\description{
Resolve an option using the dsHPC-style option chain
}
\keyword{internal}
