% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_orchestrator.R
\name{.find_mask_in_artifact}
\alias{.find_mask_in_artifact}
\title{Find a mask file within a dsHPC artifact directory}
\usage{
.find_mask_in_artifact(artifact_path, sample_id)
}
\description{
Find a mask file within a dsHPC artifact directory
}
\keyword{internal}
