% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.get_label_uri}
\alias{.get_label_uri}
\title{Get a specific label set's URI from the manifest}
\usage{
.get_label_uri(manifest, label_set_name)
}
\description{
Get a specific label set's URI from the manifest
}
\keyword{internal}
