% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.getImagingHandle}
\alias{.getImagingHandle}
\title{Retrieve an imaging handle from storage}
\usage{
.getImagingHandle(symbol)
}
\arguments{
\item{symbol}{Character; handle symbol name.}
}
\value{
The imaging handle.
}
\description{
Retrieve an imaging handle from storage
}
\keyword{internal}
