% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_options.R
\name{.imaging_analysis_home}
\alias{.imaging_analysis_home}
\title{Root directory for dsImaging analysis state.}
\usage{
.imaging_analysis_home()
}
\description{
Root directory for dsImaging analysis state.
}
\keyword{internal}
