% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_runners.R
\name{.imaging_container_config}
\alias{.imaging_container_config}
\title{Resolve optional dsImaging container image configuration.}
\usage{
.imaging_container_config(runner, args_template, command = "python")
}
\description{
Resolve optional dsImaging container image configuration.
}
\keyword{internal}
