% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_options.R
\name{.imaging_credentials_path}
\alias{.imaging_credentials_path}
\title{Path to the administrator-managed worker credential store.}
\usage{
.imaging_credentials_path()
}
\description{
Path to the administrator-managed worker credential store.
}
\keyword{internal}
