% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.imaging_handle_from_resource_client}
\alias{.imaging_handle_from_resource_client}
\title{Build a handle only from a ResourceClient bound to its Resource URL}
\usage{
.imaging_handle_from_resource_client(client, symbol)
}
\description{
Build a handle only from a ResourceClient bound to its Resource URL
}
\keyword{internal}
