% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{.imaging_list_generations_admin}
\alias{.imaging_list_generations_admin}
\title{List asset generations}
\usage{
.imaging_list_generations_admin(dataset_id = NULL, state = NULL)
}
\arguments{
\item{dataset_id}{Optional dataset id.}

\item{state}{Optional generation state.}
}
\description{
List asset generations
}
\keyword{internal}
