% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_tracking.R
\name{.imaging_set_generation_tracking_id}
\alias{.imaging_set_generation_tracking_id}
\title{Persist a root on a pre-existing generation during rolling upgrades}
\usage{
.imaging_set_generation_tracking_id(generation_id, tracking_id)
}
\description{
Persist a root on a pre-existing generation during rolling upgrades
}
\keyword{internal}
