% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_tracking.R
\name{.imaging_tracking_publish_job_asset}
\alias{.imaging_tracking_publish_job_asset}
\title{Publish the asset produced by a direct dsHPC child}
\usage{
.imaging_tracking_publish_job_asset(job_id, asset_id)
}
\description{
Publish the asset produced by a direct dsHPC child
}
\keyword{internal}
