% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privacy_contract.R
\name{.imaging_worker_manifest}
\alias{.imaging_worker_manifest}
\title{Build a worker manifest containing only request-authorized asset aliases}
\usage{
.imaging_worker_manifest(authorized, asset_names = character(0))
}
\description{
Build a worker manifest containing only request-authorized asset aliases
}
\keyword{internal}
