% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_orchestrator.R
\name{.is_transient_job_submit_error}
\alias{.is_transient_job_submit_error}
\title{Detect transient job submission errors that should be retried}
\usage{
.is_transient_job_submit_error(msg)
}
\description{
Detect transient job submission errors that should be retried
}
\keyword{internal}
