% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.manifest_for_registry_listing}
\alias{.manifest_for_registry_listing}
\title{Resolve a manifest for registry listing}
\usage{
.manifest_for_registry_listing(dataset_id, entry)
}
\description{
Resolve a manifest for registry listing
}
\keyword{internal}
