% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_install_models.R
\name{.models_dir}
\alias{.models_dir}
\title{Get the administrator-owned models directory}
\usage{
.models_dir()
}
\description{
Get the administrator-owned models directory
}
\keyword{internal}
