% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_s3.R
\name{.parse_s3_uri}
\alias{.parse_s3_uri}
\title{Parse an S3 URI into bucket and key}
\usage{
.parse_s3_uri(uri)
}
\arguments{
\item{uri}{Character; "s3://bucket/key/path"}
}
\value{
Named list: bucket, key.
}
\description{
Parse an S3 URI into bucket and key
}
\keyword{internal}
