% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_s3.R
\name{.persist_s3_credential}
\alias{.persist_s3_credential}
\title{Persist S3 credentials for worker-side dataset resolution}
\usage{
.persist_s3_credential(ref, cred)
}
\description{
Persist S3 credentials for worker-side dataset resolution
}
\keyword{internal}
