% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.resolve_backend_s3_config}
\alias{.resolve_backend_s3_config}
\title{Resolve S3 config from backend object}
\usage{
.resolve_backend_s3_config(backend)
}
\description{
Resolve S3 config from backend object
}
\keyword{internal}
