% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_s3.R
\name{.s3_base_url}
\alias{.s3_base_url}
\title{Extract base_url from endpoint (strip scheme)}
\usage{
.s3_base_url(endpoint)
}
\description{
Extract base_url from endpoint (strip scheme)
}
\keyword{internal}
