% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_s3.R
\name{.s3_head}
\alias{.s3_head}
\title{HEAD request on S3 object}
\usage{
.s3_head(config, bucket, key)
}
\value{
list(exists, size, last_modified) or NULL
}
\description{
HEAD request on S3 object
}
\keyword{internal}
