% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_client.R
\name{.validate_imaging_resource_object}
\alias{.validate_imaging_resource_object}
\title{Validate the provenance-bearing resourcer object accepted by the client}
\usage{
.validate_imaging_resource_object(resource)
}
\description{
Validate the provenance-bearing resourcer object accepted by the client
}
\keyword{internal}
