% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{.validate_uri}
\alias{.validate_uri}
\title{Validate a URI (s3:// or local absolute path)}
\usage{
.validate_uri(uri, label = "uri")
}
\description{
Validate a URI (s3:// or local absolute path)
}
\keyword{internal}
