% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_runners.R
\name{.with_container}
\alias{.with_container}
\title{Add optional container metadata to a runner configuration.}
\usage{
.with_container(config, runner, args_template, command = "python")
}
\description{
Add optional container metadata to a runner configuration.
}
\keyword{internal}
