% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_runners.R
\name{.write_runner_yaml}
\alias{.write_runner_yaml}
\title{Write a runner YAML to DSHPC_HOME/runners/}
\usage{
.write_runner_yaml(runners_dir, name, config)
}
\description{
Write a runner YAML to DSHPC_HOME/runners/
}
\keyword{internal}
