% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{get_generation}
\alias{get_generation}
\title{Get generation summary}
\usage{
get_generation(generation_id)
}
\arguments{
\item{generation_id}{Character; generation identifier.}
}
\value{
Named list or NULL.
}
\description{
Get generation summary
}
