% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{get_sample_manifest}
\alias{get_sample_manifest}
\title{Get a sample manifest}
\usage{
get_sample_manifest(dataset_id, collection_seal, sample_id)
}
\arguments{
\item{dataset_id}{Character; dataset identifier.}

\item{collection_seal}{Character; immutable collection-snapshot seal.}

\item{sample_id}{Character; sample identifier.}
}
\value{
Named list or NULL.
}
\description{
Get a sample manifest
}
