% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{imagingAssetsDS}
\alias{imagingAssetsDS}
\title{List Dataset Assets}
\usage{
imagingAssetsDS(handle_symbol)
}
\arguments{
\item{handle_symbol}{Character; symbol of the imaging handle.}
}
\value{
Data.frame with columns: name, type.
}
\description{
DataSHIELD AGGREGATE method. Returns the names and types of assets
available in the imaging dataset.
}
