% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{imagingCollectionRecoverDS}
\alias{imagingCollectionRecoverDS}
\title{Recover a collection generation}
\usage{
imagingCollectionRecoverDS(reference_symbol)
}
\arguments{
\item{reference_symbol}{Session symbol holding the collection workflow.}
}
\description{
Recover a collection generation
}
