% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{imagingCollectionStatusDS}
\alias{imagingCollectionStatusDS}
\title{Get collection status}
\usage{
imagingCollectionStatusDS(reference_symbol)
}
\arguments{
\item{reference_symbol}{Session symbol holding the collection workflow.}
}
\description{
Get collection status
}
