% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{imagingDescribeProfileDS}
\alias{imagingDescribeProfileDS}
\title{Describe a radiomics profile}
\usage{
imagingDescribeProfileDS(profile_name)
}
\arguments{
\item{profile_name}{Profile name.}
}
\description{
Describe a radiomics profile
}
