% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{imagingGetManifestDS}
\alias{imagingGetManifestDS}
\title{Get the manifest from an imaging handle}
\usage{
imagingGetManifestDS(handle_symbol)
}
\arguments{
\item{handle_symbol}{Character; the symbol name.}
}
\value{
Parsed manifest list, or NULL.
}
\description{
Get the manifest from an imaging handle
}
