% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_capabilities.R
\name{imagingListModelsDS}
\alias{imagingListModelsDS}
\title{List Installed Models (no admin needed)}
\usage{
imagingListModelsDS()
}
\value{
Data.frame of installed models.
}
\description{
DataSHIELD AGGREGATE method.
}
