% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{imagingProcessAssetWorkflowDS}
\alias{imagingProcessAssetWorkflowDS}
\title{Submit a generic clinical imaging asset workflow}
\usage{
imagingProcessAssetWorkflowDS(request_encoded)
}
\arguments{
\item{request_encoded}{B64/JSON request payload.}
}
\value{
A job handle assigned server-side.
}
\description{
Submit a generic clinical imaging asset workflow
}
