% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{imagingProcessQcCollectionDS}
\alias{imagingProcessQcCollectionDS}
\title{Submit a QC metrics workflow}
\usage{
imagingProcessQcCollectionDS(request_encoded)
}
\arguments{
\item{request_encoded}{B64/JSON request payload.}
}
\value{
A job handle assigned server-side.
}
\description{
Submit a QC metrics workflow
}
