% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{imagingProcessRadiomicsAssetDS}
\alias{imagingProcessRadiomicsAssetDS}
\title{Submit a radiomics extraction job}
\usage{
imagingProcessRadiomicsAssetDS(request_encoded)
}
\arguments{
\item{request_encoded}{B64/JSON request payload.}
}
\value{
A job handle assigned server-side.
}
\description{
Submit a radiomics extraction job
}
