% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_workflows.R
\name{imagingProcessRadiomicsCollectionDS}
\alias{imagingProcessRadiomicsCollectionDS}
\title{Process a radiomics collection}
\usage{
imagingProcessRadiomicsCollectionDS(request_encoded)
}
\arguments{
\item{request_encoded}{B64/JSON request payload.}
}
\value{
A job/generation handle assigned server-side.
}
\description{
Process a radiomics collection
}
