% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{imagingValidateDS}
\alias{imagingValidateDS}
\title{Validate Imaging Dataset}
\usage{
imagingValidateDS(handle_symbol)
}
\arguments{
\item{handle_symbol}{Character; symbol of the imaging handle.}
}
\value{
Named list with validation results.
}
\description{
DataSHIELD AGGREGATE method. Runs security and integrity checks on
the imaging dataset.
}
