% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{list_datasets}
\alias{list_datasets}
\title{List all datasets in the registry}
\usage{
list_datasets()
}
\value{
Named list of dataset entries.
}
\description{
List all datasets in the registry
}
\keyword{internal}
