% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{parse_manifest}
\alias{parse_manifest}
\title{Parse an imaging dataset manifest}
\usage{
parse_manifest(manifest_uri, backend = storage_backend("file"))
}
\arguments{
\item{manifest_uri}{Character; URI of the manifest YAML.}

\item{backend}{A dsimaging_backend object (from resolve_dataset).}
}
\value{
A validated manifest list.
}
\description{
Fetches the manifest YAML from the given backend and validates it.
}
\keyword{internal}
