% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptor.R
\name{print.ImagingDatasetDescriptor}
\alias{print.ImagingDatasetDescriptor}
\title{Print method for ImagingDatasetDescriptor}
\usage{
\method{print}{ImagingDatasetDescriptor}(x, ...)
}
\arguments{
\item{x}{An ImagingDatasetDescriptor.}

\item{...}{Ignored.}
}
\description{
Print method for ImagingDatasetDescriptor
}
