% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets_interface.R
\name{promote_asset_alias}
\alias{promote_asset_alias}
\title{Promote an Alias}
\usage{
promote_asset_alias(dataset_id, alias, asset_id)
}
\arguments{
\item{dataset_id}{Character.}

\item{alias}{Character; the alias name.}

\item{asset_id}{Character; the asset to point to.}
}
\description{
NOT a DataSHIELD method -- admin/server-side utility.
}
