% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_profiles.R
\name{read_imaging_radiomics_profile}
\alias{read_imaging_radiomics_profile}
\title{Read a radiomics profile as a list}
\usage{
read_imaging_radiomics_profile(profile_name)
}
\arguments{
\item{profile_name}{Profile name without the \code{.yaml} suffix.}
}
\value{
Parsed profile as a list.
}
\description{
Read a radiomics profile as a list
}
