% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asset_db.R
\name{update_generation}
\alias{update_generation}
\title{Update generation state}
\usage{
update_generation(generation_id, ...)
}
\arguments{
\item{generation_id}{Character; generation identifier.}

\item{...}{Named fields to update in \code{asset_generations}.}
}
\description{
Update generation state
}
