% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{validate_manifest}
\alias{validate_manifest}
\title{Validate a parsed manifest}
\usage{
validate_manifest(manifest, source = "manifest")
}
\arguments{
\item{manifest}{List; parsed manifest.}

\item{source}{Character; source URI for error messages.}
}
\description{
Validate a parsed manifest
}
\keyword{internal}
