% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_model_registry.R
\name{.get_model_config}
\alias{.get_model_config}
\title{Get a segmentation model config by name}
\usage{
.get_model_config(model_name)
}
\description{
Get a segmentation model config by name
}
\keyword{internal}
