% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend_s3.R
\name{.s3_download}
\alias{.s3_download}
\title{Download an S3 object to a local file}
\usage{
.s3_download(config, bucket, key, dest)
}
\description{
Download an S3 object to a local file
}
\keyword{internal}
