% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_client.R
\name{ImagingDatasetResourceClient}
\alias{ImagingDatasetResourceClient}
\title{ImagingDatasetResourceClient R6 class}
\description{
ImagingDatasetResourceClient R6 class

ImagingDatasetResourceClient R6 class
}
\details{
Resolves direct \code{imaging+dataset://} URLs via the registry or S3
credentials from an Opal resource. Armadillo Resources carry the strict
\code{dsimaging-dataset:<dataset_id>} locator in \code{format} and are
resolved only through the administrator-managed registry.
}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{ImagingDatasetResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImagingDatasetResourceClient-new}{\code{ImagingDatasetResourceClient$new()}}
\item \href{#method-ImagingDatasetResourceClient-getManifest}{\code{ImagingDatasetResourceClient$getManifest()}}
\item \href{#method-ImagingDatasetResourceClient-getManifestUri}{\code{ImagingDatasetResourceClient$getManifestUri()}}
\item \href{#method-ImagingDatasetResourceClient-getDatasetId}{\code{ImagingDatasetResourceClient$getDatasetId()}}
\item \href{#method-ImagingDatasetResourceClient-getBackend}{\code{ImagingDatasetResourceClient$getBackend()}}
\item \href{#method-ImagingDatasetResourceClient-asDataFrame}{\code{ImagingDatasetResourceClient$asDataFrame()}}
\item \href{#method-ImagingDatasetResourceClient-asImagingDescriptor}{\code{ImagingDatasetResourceClient$asImagingDescriptor()}}
\item \href{#method-ImagingDatasetResourceClient-clone}{\code{ImagingDatasetResourceClient$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="close"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-close'><code>resourcer::ResourceClient$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-downloadFile'><code>resourcer::ResourceClient$downloadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getConnection"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getConnection'><code>resourcer::ResourceClient$getConnection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-new}{}}}
\subsection{Method \code{new()}}{
Create and resolve an imaging dataset resource client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$new(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{Resource object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-getManifest"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-getManifest}{}}}
\subsection{Method \code{getManifest()}}{
Return the parsed imaging manifest.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$getManifest()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-getManifestUri"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-getManifestUri}{}}}
\subsection{Method \code{getManifestUri()}}{
Return the resolved manifest URI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$getManifestUri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-getDatasetId"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-getDatasetId}{}}}
\subsection{Method \code{getDatasetId()}}{
Return the dataset identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$getDatasetId()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-getBackend"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-getBackend}{}}}
\subsection{Method \code{getBackend()}}{
Return the resolved storage backend.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$getBackend()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-asDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-asDataFrame}{}}}
\subsection{Method \code{asDataFrame()}}{
Refuse generic data-frame materialization. Imaging
resources must pass the patient-level admission performed by
\code{imagingInitDS()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$asDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-asImagingDescriptor"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-asImagingDescriptor}{}}}
\subsection{Method \code{asImagingDescriptor()}}{
Convert the manifest to an imaging dataset descriptor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$asImagingDescriptor()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
