% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolver.R
\name{ImagingDatasetResourceResolver}
\alias{ImagingDatasetResourceResolver}
\title{ImagingDatasetResourceResolver R6 class}
\description{
ImagingDatasetResourceResolver R6 class

ImagingDatasetResourceResolver R6 class
}
\details{
Resolves direct \code{imaging+dataset://} URLs and Armadillo Resources whose
strict \code{format} locator is \code{dsimaging-dataset:<dataset_id>}.
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{ImagingDatasetResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImagingDatasetResourceResolver-isFor}{\code{ImagingDatasetResourceResolver$isFor()}}
\item \href{#method-ImagingDatasetResourceResolver-newClient}{\code{ImagingDatasetResourceResolver$newClient()}}
\item \href{#method-ImagingDatasetResourceResolver-clone}{\code{ImagingDatasetResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
Check whether this resolver handles a resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Resource-like object with \code{url} and optional \code{format} fields.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
Create a resource client for the resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Resource-like object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImagingDatasetResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImagingDatasetResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImagingDatasetResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
